+++ packages/web/src/app/[lng]/(auth)/memos/_utils/index.ts
@@ -1,0 +1,1 @@
+export * from './Sentry';
